move and migrate models
split models between login and stock
cleared database, but did not delete app yet
